chore(kubeconfig): force dependency on nodepool
as explain in the scaleway_k8s_doc: The null_resource is needed because
when the cluster is created, it's status is pool_required, but the
kubeconfig can already be downloaded. It leads the kubernetes provider
to start creating its objects, but the DNS entry for the Kubernetes
master is not yet ready, that's why it's needed to wait for at least a
pool.

Signed-off-by: Xavier Rakotomamonjy <xavier.rakotomamonjy@camptocamp.com>